Move and rename to MapType
