+++ cc_dynamodb3/cc_types/__init__.py
@@ -1,0 +1,1 @@
+from .types import SetType, MapType
